% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAcquisition.R
\name{bacteria.data}
\alias{bacteria.data}
\title{Query Bacteria Data from DEQ Databases}
\usage{
bacteria.data(start_date, end_date, df_stns, chr_wtsd = NULL,
  save_direct = NULL)
}
\arguments{
\item{start_date, }{starting date of the form "'1990-01-01 00:00:00'"}

\item{end_date, }{ending date of the form "'1990-01-01 00:00:00'"}

\item{df_stns, }{dataframe of monitoring stations 'station', watershed 'watershed', and gage numbers 'gage'}

\item{chr_wtsd, }{the watershed of interest, ie. "Upper Yaquina River". Only need if saving into a directory ('save_direct'). Default is NULL.}

\item{save_direct, }{the directory to save files, default is NULL}
}
\value{
bacteria.data dataframe of queried data
}
\description{
This function will allow users to acquire necessary data from the DEQ databases to create Load Duration Curves
}
