% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_annual_ldc.R
\name{calc_annual_ldc}
\alias{calc_annual_ldc}
\title{Calculate annualized load duration curve}
\usage{
calc_annual_ldc(
  .tbl,
  Q = NULL,
  C = NULL,
  Date = NULL,
  allowable_concentration = NULL,
  breaks = c(1, 0.8, 0.4, 0),
  labels = c("High Flows", "Medium Flows", "Low Flows"),
  estimator = 6,
  n = 500
)
}
\arguments{
\item{.tbl}{data frame with at least three columns Q (discharge or flow), C
(associated pollutant concentration), and Date.}

\item{Q}{variable name in .tbl for discharge or flow. This must be of class
`units`, typically with a units value of "ft^3/s".}

\item{C}{variable name in .tbl for associated pollutant concentration at a
given flow value. This must be of class `units`, typically with a units
value of "mg/L" or "cfu/100mL".}

\item{Date}{variable name in .tbl for the event Date. This variable must be
of class `Date`.}

\item{allowable_concentration}{an object of class \code{units} specifying the
allowable pollutant concentration.}

\item{breaks}{a numeric vector of break points for flow categories. Must be
of length of labels + 1. defaults to \code{c(1, 0.8, 0.4, 0)}.}

\item{labels}{labels for the categories specified by breaks.}

\item{estimator}{numeric, one of \code{c(5,6,7,8,9)}. \code{6} is the default
method correponding to the Weibull plotting position. Further details are
provided in \code{stats::quantile()}.}

\item{n}{numeric, the length of generated probability points. Larger n may
result in a slightly smoother curve at a cost of increased processing time.
The probability points are used to generate the continuous sample quantiles
types 5 to 9 (see \code{quantile}).}
}
\value{
list of two tibbles (Q and C). Includes variables in .tbl and
  Daily_Flow_Volume (discharge volume), Daily_Load (pollutant sample volume),
  P_Exceedance (exeedance probability), Flow_Category (as defined by breaks
  and labels).
}
\description{
Calculates the median annual ldc with confidence intervals.
}
