% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_ldc.R
\name{draw_ldc}
\alias{draw_ldc}
\title{Draw a load duration curve}
\usage{
draw_ldc(
  .tbl_calc,
  .tbl_summ,
  y_lab = NULL,
  ldc_legend_name = "Allowable Load at State Water Quality Standard",
  measurement_name = "Measurement Value",
  measurement_shape = 21,
  measurement_color = "dodgerblue",
  measurement_alpha = 1,
  summary_name = "Summarized Measured Load",
  summary_stat_shape = 12,
  summary_stat_color = "red",
  label_nudge_y = 0,
  label_font_family = "Arial",
  label_font_size = 3,
  label_break = TRUE
)
}
\arguments{
\item{.tbl_calc}{data frame object}

\item{.tbl_summ}{data frame object}

\item{y_lab}{optional string for y-axis label name, will be appended with units automatically. default is NULL.}

\item{ldc_legend_name}{string, provides the name used for the allowable pollutant load line in the legend. required.}

\item{measurement_name}{string, provides the name used for measured load values in the legend. required.}

\item{measurement_shape}{aesthetic value passed to the layer plotting measured load values. defaults to \code{21}.}

\item{measurement_color}{aesthetic value passed to the layer plotting measured load values. defaults to \code{"dodgerblue"}.}

\item{measurement_alpha}{aesthetic value passed to the layer plotting measured load values. defaults to \code{1}.}

\item{summary_name}{string, provides the name used for summary statistic values in the legend. required.}

\item{summary_stat_shape}{aesthetic value passed to the layer plotting summary statistic values. defaults to \code{12}.}

\item{summary_stat_color}{aesthetic value passed to the layer plotting summary statistic values. defaults to \code{"red"}.}

\item{label_nudge_y}{numeric value to vertically nudge flow category labels. If a log10 transformed scale is being used, a log value is probably appropriate for example \code{log10(1000)}.}

\item{label_font_family}{string specifying font family to use in flow category labels.}

\item{label_font_size}{numeric value specifying font size to use in flow category labels.}

\item{label_break}{logical, add line breaks to flow cateogry labels. Labels will break at spaces.}
}
\value{
ggplot object
}
\description{
Creates a load duration curve visualization from the outputs of \code{calc_ldc} and \code{summ_ldc} as a ggplot object.
}
