% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FdcSSEstimate.R
\name{fdc.ss.estimate}
\alias{fdc.ss.estimate}
\title{Flow duration curve estimate values}
\usage{
fdc.ss.estimate(ss.fn = NULL, ss.path = NULL, Eq.Region = 1)
}
\arguments{
\item{ss.fn, }{paste(st, stationID, .xml)}

\item{ss.path, }{path for streamstats xml (use getPath to find the correct directory, ie. getPath('StreamStatsBacteria'))}

\item{Eq.Region, }{default = 1}
}
\value{
st.ss.est dataframe of flow duration estimates
}
\description{
This function will output a dataframe of flow duration estimates based on the Stream Stats XML file and ecoregion (???). However, this function is only set up for Ecoregion 1.
}
