% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPath.R
\name{getPath}
\alias{getPath}
\title{Path Directory for working on Load Duration Curves}
\usage{
getPath(folder = NULL, wtsd = NULL)
}
\arguments{
\item{folder, }{directory folder name; default NULL}

\item{wtsd, }{watershed; default NULL}
}
\value{
path specific to watershed and folder inputs
}
\description{
This function will help users organize files into saved folders specific to watersheds and data collected
}
