% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldcPlot.R
\name{ldc.plot}
\alias{ldc.plot}
\title{Load Duration Curve Plot}
\usage{
ldc.plot(ldc.stn, ldc.crit, y.lims = NULL)
}
\arguments{
\item{ldc.stn, }{list of 4 containing $points (Date, ecoli, digression, reduction, flow, Vol.day, flow.exceed, flow.zone, and load.day);
$summary.points (N, Digressions, Max.Reduction, Max.Reduction.FLZ, Max.Reduction.flow.exceed, Max.Reduction.ecoli);
$geomns (flow.exceed, Vol.day, load.day, conc, N, digression, reduction, desc);
$summary.geomns (N, Digressions, Max.Reducition, Max.Reduction.FLZ, Max.Reduction.conc)}

\item{ldc.crit, }{list of 2 containing $ldc.max (Date, flow, vol.day, flow.exceed, load.day);
$ldc.gmn (flow.exceed, Vol.day, load.day)}

\item{y.lims, }{initial LDC limits [c(1e+15,1E-05)]}
}
\value{
Load Duration Curve plot
}
\description{
This function will output a load duration curve plot for a single monitoring station
}
