% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldcCriteriaEcoli.R
\name{ldcCriteriaEcoli}
\alias{ldcCriteriaEcoli}
\title{Load Duration Curve Criteria -- List of 2}
\usage{
ldcCriteriaEcoli(flow.data)
}
\arguments{
\item{flow.data, }{dataframe containing station, date (YYYY-MM-DD), flow_cfs}
}
\value{
List of 2 containing $ldc.max and $ldc.gmn
}
\description{
This function will create a list of two containing: 1) ldc.max (Date, flow, Vol.day, flow.exceed, load.day), 2) ldc.gmn (flow.exceed, Vol.day, load.day). This is an input for ldcPlot and scatterPlot
}
