% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterBacteriaTSPlot}
\alias{scatterBacteriaTSPlot}
\title{Scatter Bacteria Time Series Plot}
\usage{
scatterBacteriaTSPlot(ldc, ldc.crit, est.flow, plot.fn = NULL)
}
\arguments{
\item{ldc, }{list of 4 containing $points (Date, ecoli, digression, reduction, flow, Vol.day, flow.exceed, flow.zone, and load.day);
$summary.points (N, Digressions, Max.Reduction, Max.Reduction.FLZ, Max.Reduction.flow.exceed, Max.Reduction.ecoli);
$geomns (flow.exceed, Vol.day, load.day, conc, N, digression, reduction, desc);
$summary.geomns (N, Digressions, Max.Reducition, Max.Reduction.FLZ, Max.Reduction.conc)}

\item{ldc.crit, }{list of 2 containing $ldc.max (Date, flow, vol.day, flow.exceed, load.day);
$ldc.gmn (flow.exceed, Vol.day, load.day)}

\item{est.flow, }{dataframe including date, value, flow.exceed}

\item{plot.fn, }{filepath for the scatterplot to be saved in. Default is NULL.}
}
\value{
bacteria time series scatter plot
}
\description{
This function will create a scatter plot of bacteria data over time
}
