% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_ldc.R
\name{summ_ldc}
\alias{summ_ldc}
\title{Summarize load duration curve}
\usage{
summ_ldc(.tbl, Q, C, Exceedance, groups, method = "geomean")
}
\arguments{
\item{.tbl}{data frame, prefferably the output from \code{calc_ldc()}.}

\item{Q}{variable name in .tbl for discharge or flow. This must have unit set, typically "ft^3/s".}

\item{C}{variable name in .tbl for associated pollutant concentration at a given flow value. This must have a unit set, typically "mg/L" or "cfu/100mL".}

\item{Exceedance}{variable name in .tbl with flow/load exceedance probabilities.}

\item{groups}{variable name in .tbl with categorized flow names.}

\item{method}{string that describes the summary statistic used for the pollutant concentration. Must be one of \code{c('geomean', 'mean', 'median')}.}
}
\value{
a data frame
}
\description{
Summarize load duration curve
}
