% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tres_palacios}
\alias{tres_palacios}
\title{Mean daily flow and point E. coli bacteria measurements.}
\format{
A data frame with 7671 rows and 4 variables:
\describe{
  \item{site_no}{USGS gage number}
  \item{Date}{Observation Date}
  \item{Flow}{Mean Daily Flow in cfs}
  \item{Indicator_Bacteria}{Bacteria concentration measured on the given day in MPN/100mL}
   }
}
\source{
USGS NWIS \url{https://waterdata.usgs.gov/nwis} and TCEQ SWQM {https://www.tceq.texas.gov/waterquality/monitoring}
}
\usage{
tres_palacios
}
\description{
A dataset containing the mean daily flow and E. coli bacteria concentrations
on the Tres Palacios River from 2000 through 2020.
}
\keyword{datasets}
